Jalousie, Musik und Thermostat vollständig mit Datenbank verknüpft, LightSettingsActivity teilweise

+++ app/src/main/java/com/sophiafema/home_easylife/view/VolumePicker.java
@@ -31,7 +31,7 @@ public class VolumePicker extends Picker {
 
     @Override
     public void setTextFromValue(float value) {
-        setCenterText("" + calculatePercentWithValue(value));
+        setCenterText("" + (int) calculatePercentWithValue(value));
     }
 
 }
